+++ contracts-metropolis/src/interfaces/IBaseVault.sol
@@ -122,7 +122,7 @@ interface IBaseVault is IMinimalVault, IERC20Upgradeable {
 
     function getAumAnnualFee() external view returns (uint256);
 
-    function getRange() external view returns (int32 low, int32 upper);
+    function getRange() external view returns (uint24 low, uint24 upper);
 
     function getOperators()
         external
